+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-﻿# Changelog
+## [0.4] - 2025-09-28
+### Added
+- Animowane trasy statków na mapie (różne kolory).
+- Tooltipy z czasem podróży i typem statku.
+- Animowane zakładki w panelu Toolkita.
+- Dodano katalog assets/ z grafikami.
+# Changelog
 
 ## [0.1] - (data generowania)
-- Dodano ikonÄ™ â€žToolkitâ€ť w menu
-- Utworzono panel z trzema zakĹ‚adkami (puste)
+- Dodano ikonę „Toolkit” w menu
+- Utworzono panel z trzema zakładkami (puste)
 ## [0.2] - 2025-09-28
 - Aktualizacja wersji 0.2